read msig and create .teal
